Named snapshot on clone, remove snapshot on delete
When creating a linked clone, manually create a snapshot with a unique
name to link the clone. This prevents the generic "Clone" name from
being used. We use the basename of the destination directory as it will
be a unique UUID. The name of the snapshot and the source of the vmx
file the link clone originated are stored in vm directory.

When deleting the guest, the vm directory is checked for files
containing the original vmx path and the snapshot name. If they exist,
the snapshot will be deleted before the vm directory contents are
destroyed.