Fix non-scsi primary disk resizing
Primary disk was restricted to the first scsi device. This has been
updated to cycle through sata, scsi, and ide types to locate the first
available connected device.

Check is also performed to determine if the guest is a linked clone. If
it is, growing of the primary disk will be ignored and a warning will be
emitted.