Check for existing lease entries when loading
It is possible to create multiple guests with the same MAC address which
results in multiple valid lease entries with different IP addresses and
no valid way to identify their owner. When multiple entries are encountered,
remove any entries from the list and do not allow additional matching
entries to be added.