Default to professional license when license data is empty
Reports of the new free licenses show the license check produces
empty content. If empty content is encountered, just default to
marking the installation as professional.

Note that forcing the license to always be professional is not
done here to ensure older versions continue to work as expected
and don't attempt to use features which are not accessible.